git-svn-id: https://casparcg.svn.sourceforge.net/svnroot/casparcg/server/branches/2.0 concrt-experimental@1532 362d55ac-95cf-4e76-9f9a-cbaa9c17b72d

+++ core/consumer/output.cpp
@@ -113,7 +113,7 @@ public:
 	{	
 		const auto frame = element.first;
 
-		if(frame->image_data().size() != format_desc_.size)
+		if(frame->image_data().size() != static_cast<int>(format_desc_.size))
 			return;
 
 		{
